Fix offenses

+++ src/Kobo/ImageProcessor/CoverTransformer.php
@@ -94,11 +94,11 @@ class CoverTransformer
         }
 
         // Calculate the position to center the image on the canvas
-        $x = ($maxWidth - $newWidth) / 2;
-        $y = ($maxHeight - $newHeight) / 2;
+        $x = (int) (($maxWidth - $newWidth) / 2);
+        $y = (int) (($maxHeight - $newHeight) / 2);
 
         // Copy the original image onto the new image and center it
-        imagecopyresampled($image, $originalImage, (int) $x, (int) $y, 0, 0, $newWidth, $newHeight, $originalWidth, $originalHeight);
+        imagecopyresampled($image, $originalImage, $x, $y, 0, 0, (int) $newWidth, $newHeight, $originalWidth, $originalHeight);
 
         if ($grayscale) {
             imagefilter($image, IMG_FILTER_GRAYSCALE);
